[!] Fixed and made MEMCPY_TEST platform independent [!] Small bug fix when decrypting PifRam
